Moved discretize methods into separate file
