+++ teacher/fuzzy/__init__.py
@@ -1,21 +1,13 @@
-from ._base import (get_equal_width_division,
-                    get_equal_freq_division,
-                    get_fuzzy_points,
-                    get_fuzzy_triangle,
-                    fuzzy_entropy,
-                    weighted_fuzzy_entropy,
+from ._base import (get_fuzzy_points,
                     dataset_membership,
                     get_fuzzy_variables)
 from .fuzzy_set import FuzzyDiscreteSet, FuzzyContinuousSet
 from .fuzzy_variable import FuzzyVariable
+from ._discretize import fuzzy_entropy
 
 __all__ = [
-    "get_equal_width_division",
-    "get_equal_freq_division",
-    "get_fuzzy_triangle",
     "fuzzy_entropy",
     "get_fuzzy_points",
-    "weighted_fuzzy_entropy",
     "dataset_membership",
     "get_fuzzy_variables",
     "FuzzyDiscreteSet",
